@@ -1,18 +1,18 @@
 name: jelly_anim
-description: A widget for play with jelly animation in your app.
-version: 0.0.5
+description: This package contains jelly animation. Users can dynamically modify the jelly size, viewport size, jelly coordinates, overflow flag and also customize it's as per user requirements.
+version: 0.0.6-nullsafety
 homepage: https://github.com/Mindinventory/jelly_anim
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  angles: ^1.2.1
-  bezier: ^1.1.2
-  random_color: ^1.0.3
+  angles: ^2.1.0
+  bezier: ^1.2.0
+  random_color: ^1.0.6-nullsafety
 dev_dependencies:
   flutter_test:
     sdk: flutter
