@@ -22,7 +22,7 @@ class JellyAnim extends StatefulWidget {
 
   /// inside the [viewPortSize] defined size the jelly will draw.
   /// NOTE: by default the size of viewport is same as screen size.
-  final Size viewPortSize;
+  final Size? viewPortSize;
 
   /// [colors] list of colors for transition if user will not give any color then it'll take random color.
   final List<Color> colors;
@@ -32,22 +32,23 @@ class JellyAnim extends StatefulWidget {
   final JellyPosition jellyPosition;
 
   /// [fillPaint] to give custom paint values like stroke, strokeWidth, style, etc.
-  final Paint fillPaint;
+  final Paint? fillPaint;
 
   /// [allowOverFlow] flag will give us a provision to render animation outside the viewport.
   final bool allowOverFlow;
 
-  JellyAnim(
-      {@required this.radius,
-      this.allowOverFlow = false,
-      this.duration = const Duration(seconds: 5),
-      this.colors,
-      this.viewPortSize,
-      this.jellyCoordinates = 5,
-      this.jellyPosition = JellyPosition.center,
-      this.jellyCount=1,
-      this.fillPaint}){
-   assert(radius>2);
+  JellyAnim({
+    required this.radius,
+    this.allowOverFlow = false,
+    this.duration = const Duration(seconds: 5),
+    this.colors = const [],
+    this.viewPortSize,
+    this.jellyCoordinates = 5,
+    this.jellyPosition = JellyPosition.center,
+    this.jellyCount = 1,
+    this.fillPaint,
+  }) {
+    assert(radius > 2);
   }
 
   @override
@@ -57,9 +58,9 @@ class JellyAnim extends StatefulWidget {
 }
 
 class _JellyAnim extends State<JellyAnim> with TickerProviderStateMixin {
-  List<JellyConfiguration> jellyConfigurations;
-  Size size;
-  AnimationController _controller;
+  late List<JellyConfiguration> jellyConfigurations;
+  late Size size;
+  late AnimationController _controller;
   JellyPosition position = JellyPosition.center;
 
   @override
@@ -69,7 +70,7 @@ class _JellyAnim extends State<JellyAnim> with TickerProviderStateMixin {
       vsync: this,
       duration: widget.duration,
     );
-   _controller.repeat();
+    _controller.repeat();
     _controller.addStatusListener((status) {
       if (status == AnimationStatus.dismissed) {
         _controller.repeat();
@@ -77,18 +78,23 @@ class _JellyAnim extends State<JellyAnim> with TickerProviderStateMixin {
     });
   }
 
+  /// This method will draw the jelly & overlap the jelly based on value of [jellyCount].
   List<JellyConfiguration> createJellies(
       BuildContext context, int jellyCount, Size size) {
-    List<JellyConfiguration> jellyConfigurations = List();
+    List<JellyConfiguration> jellyConfigurations = [];
     for (int i = 0; i < jellyCount; i++) {
-      jellyConfigurations.add(JellyConfiguration(size,
+      jellyConfigurations.add(
+        JellyConfiguration(
+          size,
           jellyPosition: i,
           reductionRadiusFactor: 1.5 - ((i + 1) / 20),
           fillPaint: widget.fillPaint,
           jellyCoordinates: widget.jellyCoordinates,
           jellyPositionEnum: position,
           radius: widget.radius,
-          colorList: widget.colors));
+          colorList: widget.colors,
+        ),
+      );
     }
     return jellyConfigurations;
   }
@@ -102,55 +108,90 @@ class _JellyAnim extends State<JellyAnim> with TickerProviderStateMixin {
   @override
   Widget build(BuildContext context) {
     size = widget.viewPortSize ??
-        Size(MediaQuery.of(context).size.width,
-            MediaQuery.of(context).size.height);
+        Size(
+          MediaQuery.of(context).size.width,
+          MediaQuery.of(context).size.height,
+        );
 
     jellyConfigurations = createJellies(
-        context, widget.jellyCount, getJellySize(widget.jellyPosition));
+      context,
+      widget.jellyCount,
+      getJellySize(
+        widget.jellyPosition,
+      ),
+    );
     return SafeArea(
       child: AnimatedBuilder(
-          animation: _controller,
-          builder: (BuildContext context, Widget child) {
-            return CustomPaint(
-                size: size,
-                // size: getJellySize(JellyPosition.center),
-                painter: JellyPaint(
-                  allowOverFlow: widget.allowOverFlow,
-                  animation: _controller,
-                  jellyConfigurations: jellyConfigurations,
-                ));
-          }),
+        animation: _controller,
+        builder: (BuildContext context, Widget? child) {
+          return CustomPaint(
+            size: size,
+            painter: JellyPaint(
+              allowOverFlow: widget.allowOverFlow,
+              animation: _controller,
+              jellyConfigurations: jellyConfigurations,
+            ),
+          );
+        },
+      ),
     );
   }
 
+  /// This method will draw the jelly on the desired position on viewport
+  /// There is 1 Enum [JellyPosition] which contains all the positions & based
+  /// on that position we are calculating the size of jelly to that position.
   Size getJellySize(JellyPosition position) {
     switch (position) {
       case JellyPosition.bottomLeft:
-        return Size((size.width / widget.radius) + widget.radius * 2,
-            size.height * 2 - (widget.radius * 2));
+        return Size(
+          (size.width / widget.radius) + widget.radius * 2,
+          size.height * 2 - (widget.radius * 2),
+        );
       case JellyPosition.bottomCenter:
-        return Size(size.width, size.height * 2 - (widget.radius * 2));
+        return Size(
+          size.width,
+          size.height * 2 - (widget.radius * 2),
+        );
       case JellyPosition.bottomRight:
-        return Size(size.width * 2 - (widget.radius * 2) - widget.radius,
-            size.height * 1.9 - widget.radius);
+        return Size(
+          size.width * 2 - (widget.radius * 2) - widget.radius,
+          size.height * 1.9 - widget.radius,
+        );
       case JellyPosition.topRight:
-        return Size(size.width * 2 - (widget.radius * 2) - widget.radius, 80);
+        return Size(
+          size.width * 2 - (widget.radius * 2) - widget.radius,
+          80,
+        );
       case JellyPosition.topCenter:
-        return Size(size.width, size.height / 2 - (widget.radius / 2));
+        return Size(
+          size.width,
+          size.height / 2 - (widget.radius / 2),
+        );
       case JellyPosition.topLeft:
-        return Size(size.width / 5, size.height / 2 - (widget.radius * 2));
+        return Size(
+          size.width / 5,
+          size.height / 2 - (widget.radius * 2),
+        );
       case JellyPosition.centerLeft:
         return Size(
-            (size.width / widget.radius) + widget.radius * 2, size.height);
+          (size.width / widget.radius) + widget.radius * 2,
+          size.height,
+        );
       case JellyPosition.centerRight:
         return Size(
-            size.width * 2 - (widget.radius * 2) - widget.radius, size.height);
+          size.width * 2 - (widget.radius * 2) - widget.radius,
+          size.height,
+        );
       default:
-        return Size(size.width, size.height);
+        return Size(
+          size.width,
+          size.height,
+        );
     }
   }
 }
 
+/// This enum defines the position of jelly.
 enum JellyPosition {
   centerLeft,
   centerRight,
