@@ -50,238 +50,234 @@ class _MyHomePage extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: SingleChildScrollView(
-        child: Column(children: [
-          Stack(
-            children: [
-              Align(
-                alignment: Alignment.bottomLeft,
-                child: JellyAnim(
-                  jellyCount: jellyCount,
-                  radius: 20,
-                  viewPortSize: Size(200, 600),
-                  jellyCoordinates: 10,
-                  allowOverFlow: true,
-                  duration: Duration(milliseconds: 200),
-                  jellyPosition: JellyPosition.bottomCenter,
-                ),
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 120,
-                viewPortSize: Size(700, 400),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.centerLeft,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 70,
-                viewPortSize: Size(800, 700),
-                jellyCoordinates: 19,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.bottomLeft,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 120,
-                viewPortSize: Size(800, 400),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.bottomRight,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 50,
-                viewPortSize: Size(200, 100),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.bottomLeft,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 200,
-                viewPortSize: Size(400, 700),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.bottomCenter,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 800,
-                viewPortSize: Size(300, 700),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.bottomRight,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 400,
-                viewPortSize: Size(400, 500),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.centerLeft,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 350,
+      body: Stack(
+        children: [
+          Align(
+            alignment: Alignment.bottomLeft,
+            child: JellyAnim(
+              jellyCount: jellyCount,
+              radius: 20,
+              viewPortSize: Size(200, 600),
+              jellyCoordinates: 10,
+              allowOverFlow: true,
+              duration: Duration(milliseconds: 200),
+              jellyPosition: JellyPosition.bottomCenter,
+            ),
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 120,
+            viewPortSize: Size(700, 400),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.centerLeft,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 70,
+            viewPortSize: Size(800, 700),
+            jellyCoordinates: 19,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.bottomLeft,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 120,
+            viewPortSize: Size(800, 400),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.bottomRight,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 50,
+            viewPortSize: Size(200, 100),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.bottomLeft,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 200,
+            viewPortSize: Size(400, 700),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.bottomCenter,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 800,
+            viewPortSize: Size(300, 700),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.bottomRight,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 400,
+            viewPortSize: Size(400, 500),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.centerLeft,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 350,
 
-                viewPortSize: Size(400, 400),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.topRight,
-              ),
-              JellyAnim(
+            viewPortSize: Size(400, 400),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.topRight,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 450,
+
+            viewPortSize: Size(300, 800),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.bottomCenter,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 100,
+
+            viewPortSize: Size(350, 800),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.bottomLeft,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 168,
+            viewPortSize: Size(300, 300),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.centerLeft,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 200,
+            viewPortSize: Size(30, 700),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.centerRight,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 164,
+            viewPortSize: Size(700, 800),
+            jellyCoordinates: 10,
+
+            allowOverFlow: true,
+            jellyPosition: JellyPosition.bottomCenter,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 176,
+            viewPortSize: Size(300, 760),
+            jellyCoordinates: 10,
+            allowOverFlow: true,
+
+            jellyPosition: JellyPosition.center,
+          ),
+          Align(
+            alignment: Alignment.bottomRight,
+            child: JellyAnim(
+              jellyCount: jellyCount,
+              radius: 134,
+              viewPortSize: Size(300, 800),
+              jellyCoordinates: 10,
+              allowOverFlow: true,
+
+              jellyPosition: JellyPosition.center,
+            ),
+          ),
+          Align(
+            alignment: Alignment.bottomRight,
+            child: JellyAnim(
+              jellyCount: jellyCount,
+              radius: 50,
+              viewPortSize: Size(800, 600),
+              jellyCoordinates: 16,
+              allowOverFlow: true,
+
+              jellyPosition: JellyPosition.bottomCenter,
+            ),
+          ),
+          Align(
+            alignment: Alignment.topLeft,
+            child: JellyAnim(
+              jellyCount: jellyCount,
+              radius: 90,
+              viewPortSize: Size(500, 800),
+              jellyCoordinates: 26,
+              allowOverFlow: true,
+
+              jellyPosition: JellyPosition.topRight,
+            ),
+          ),
+          Align(
+            alignment: Alignment.topCenter,
+            child: JellyAnim(
+              jellyCount: jellyCount,
+              radius: 20,
+              viewPortSize: Size(350, 870),
+              jellyCoordinates: 26,
+
+              allowOverFlow: true,
+              jellyPosition: JellyPosition.bottomLeft,
+            ),
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 190,
+            viewPortSize: Size(250, 800),
+            jellyCoordinates: 7,
+            allowOverFlow: true,
+
+            jellyPosition: JellyPosition.bottomCenter,
+          ),
+          JellyAnim(
+            jellyCount: jellyCount,
+            radius: 40,
+            viewPortSize: size,
+            jellyCoordinates: 16,
+            allowOverFlow: true,
+
+            jellyPosition: JellyPosition.topLeft,
+          ),
+          Align(
+            alignment: Alignment.center,
+            child: Opacity(
+              opacity: 0.6,
+              child: JellyAnim(
                 jellyCount: jellyCount,
-                radius: 450,
-
-                viewPortSize: Size(300, 800),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.bottomCenter,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 100,
-
-                viewPortSize: Size(350, 800),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.bottomLeft,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 168,
-                viewPortSize: Size(300, 300),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.centerLeft,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 200,
-                viewPortSize: Size(30, 700),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.centerRight,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 164,
-                viewPortSize: Size(700, 800),
-                jellyCoordinates: 10,
-
-                allowOverFlow: true,
-                jellyPosition: JellyPosition.bottomCenter,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 176,
-                viewPortSize: Size(300, 760),
-                jellyCoordinates: 10,
-                allowOverFlow: true,
-
-                jellyPosition: JellyPosition.center,
-              ),
-              Align(
-                alignment: Alignment.bottomRight,
-                child: JellyAnim(
-                  jellyCount: jellyCount,
-                  radius: 134,
-                  viewPortSize: Size(300, 800),
-                  jellyCoordinates: 10,
-                  allowOverFlow: true,
-
-                  jellyPosition: JellyPosition.center,
-                ),
-              ),
-              Align(
-                alignment: Alignment.bottomRight,
-                child: JellyAnim(
-                  jellyCount: jellyCount,
-                  radius: 50,
-                  viewPortSize: Size(800, 600),
-                  jellyCoordinates: 16,
-                  allowOverFlow: true,
-
-                  jellyPosition: JellyPosition.bottomCenter,
-                ),
-              ),
-              Align(
-                alignment: Alignment.topLeft,
-                child: JellyAnim(
-                  jellyCount: jellyCount,
-                  radius: 90,
-                  viewPortSize: Size(500, 800),
-                  jellyCoordinates: 26,
-                  allowOverFlow: true,
-
-                  jellyPosition: JellyPosition.topRight,
-                ),
-              ),
-              Align(
-                alignment: Alignment.topCenter,
-                child: JellyAnim(
-                  jellyCount: jellyCount,
-                  radius: 20,
-                  viewPortSize: Size(350, 870),
-                  jellyCoordinates: 26,
-
-                  allowOverFlow: true,
-                  jellyPosition: JellyPosition.bottomLeft,
-                ),
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 190,
-                viewPortSize: Size(250, 800),
-                jellyCoordinates: 7,
-                allowOverFlow: true,
-
-                jellyPosition: JellyPosition.bottomCenter,
-              ),
-              JellyAnim(
-                jellyCount: jellyCount,
-                radius: 40,
+                radius: 150,
                 viewPortSize: size,
+
                 jellyCoordinates: 16,
                 allowOverFlow: true,
-
-                jellyPosition: JellyPosition.topLeft,
+                jellyPosition: JellyPosition.center,
               ),
-              Align(
-                alignment: Alignment.center,
-                child: Opacity(
-                  opacity: 0.6,
-                  child: JellyAnim(
-                    jellyCount: jellyCount,
-                    radius: 150,
-                    viewPortSize: size,
-
-                    jellyCoordinates: 16,
-                    allowOverFlow: true,
-                    jellyPosition: JellyPosition.center,
-                  ),
-                ),
-              ),
-              SizedBox(
-                height: MediaQuery.of(context).size.height,
-                width: MediaQuery.of(context).size.width,
-                child: Center(
-                  child: Text(
-                    "JELLY",
-                    textAlign: TextAlign.center,
-                    style: GoogleFonts.varelaRound(
-                        letterSpacing: 20,
-                        color: Colors.white70.withOpacity(0.5),
-                        fontSize: 60,
-                        fontWeight: FontWeight.w800),
-                  ),
-                ),
-              ),
-            ],
+            ),
           ),
-        ]),
+          SizedBox(
+            height: MediaQuery.of(context).size.height,
+            width: MediaQuery.of(context).size.width,
+            child: Center(
+              child: Text(
+                "JELLY",
+                textAlign: TextAlign.center,
+                style: GoogleFonts.varelaRound(
+                    letterSpacing: 20,
+                    color: Colors.white70.withOpacity(0.5),
+                    fontSize: 60,
+                    fontWeight: FontWeight.w800),
+              ),
+            ),
+          ),
+        ],
       ),
     );
     // return SafeArea(
